pyproject.toml
[project]
name = "kr-email-health"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "dnspython",
    "requests",
    "python-dotenv",
    "fastapi[standard]>=0.115",
    "psycopg[binary]>=3.1",
    "apscheduler>=3.10",
    "resend>=2.0",
    "typer>=0.9.0",
    "bcrypt>=4.0",
    "itsdangerous>=2.0",
    "defusedxml>=0.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
pdf = ["weasyprint>=60"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
senderfit = "check:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
